[docs] 회원가입 폴더 생성
