+++ src/App.js
@@ -7,8 +7,8 @@ import InterestPage from "./pages/InterestPage";
 import Mypage from "./pages/Mypage";
 import ProfilePage from "./pages/ProfilePage";
 import SignupPage from "./pages/SignupPage.jsx";
-import Signup from "./components/Signup.jsx";
-import Address from "./components/Address";
+import Signup from "./components/signup/Signup.jsx";
+import Address from "./components/signup/Address";
 import CompletePage from "./pages/CompletePage";
 import MainPage from "./pages/MainPage";
 import CommunityPage from "./pages/CommunityPage";
