--- src/components/Navbar.jsx
@@ -1,1 +1,0 @@
-const Navbar = () => {};
